De-raced launcher and linter CLI.

+++ bin/cli.js
@@ -9,30 +9,46 @@ const child_process = require('child_process');
 
 const options = minimist(process.argv.slice(2));
 
-const project = require(path.resolve(__dirname + '/../package.json'));
+const debounceTimeout = parseInt(options.debounce || 250);
 
+const project = require(path.resolve(__dirname + '/../package.json'));
 const athenaCommand = path.resolve(__dirname + '/cli.js');
+
+let paused = false;
 let athenaProcess;
 
-function spawnAthena() {
+function launchAthena() {
   try {
     if (options.lint) {
       console.log('Linting...');
-      const output = child_process.execSync(project.scripts['api:lint'] + ' --fix');
-      console.log(output.toString());
+      child_process.execSync(project.scripts['api:lint'] + ' --fix', {
+        stdio: 'inherit'
+      });
+    }
+
+    if (athenaProcess) {
+      console.log('Reloading Athena...');
+    } else {
+      console.log('Launching Athena...');
     }
 
     const args = options._;
     if (options.config) {
-      args.push(`--config=${options.config}`);
+      args.push(`--config=${ options.config }`);
     }
 
-    athenaProcess = child_process.spawn(athenaCommand, args, {
+    athenaProcess = child_process.fork(athenaCommand, args, {
       detached: false,
       stdio: 'inherit'
     });
+
+    athenaProcess.on('message', function() {
+      paused = false;
+    });
+
   } catch (error) {
-    console.log('Error spawning Athena', error);
+    console.log('Error launching Athena:\n  ', error.message);
+    paused = false;
   }
 }
 
@@ -44,24 +60,30 @@ if (options.help) {
   const watchDirectory = path.resolve(__dirname + '/../server/');
   let debounce = 0;
 
-  spawnAthena();
+  launchAthena();
 
   fs.watch(watchDirectory, {
     persistent: true,
     recursive: true
   }, function (eventType, filename) {
-    if (!filename.startsWith('.') && filename.endsWith('.js')) {
+    if (!paused && !filename.startsWith('.') && filename.endsWith('.js')) {
       if (debounce) {
         clearTimeout(debounce);
       }
 
       debounce = setTimeout(function () {
+        paused = true;
+
         if (athenaProcess) {
+          athenaProcess.on('exit', function() {
+            launchAthena();
+          });
+
           athenaProcess.kill();
+        } else {
+          launchAthena();
         }
-        console.log('Reloading Athena...');
-        spawnAthena();
-      }, 250);
+      }, debounceTimeout);
     }
   });
 } else {
@@ -72,5 +94,8 @@ if (options.help) {
 
   const Athena = require('../server/athena');
   const athena = new Athena(config);
-  athena.boot();
+
+  athena.boot(function() {
+    process.send('ready');
+  });
 }
